adding the second task

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	if (array != NULL && action != NULL)
 	{
-		long unsigned int i;
+		size_t i;
 
 		for (i = 0; i < size; i++)
 		{
